+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `From` function to generate identifiers from one or more string values.
+
 ## [0.7.0] - 2025-10-09
 
 ### Changed